Initial Commit
I had done lots of work with Git as VCS, but partway through there was
an error and it had something to do with PhpStorm.  I couldn't figure it
out and so I decided to create a new repository.  So much wasted time.
Didn't learn how to solve my problem :(